backend/requirements.txt
fastapi
uvicorn
uvicorn[standard]
websockets
pydantic[email]
google-genai
python-dotenv